Separate message providers for search and login
