+++ src/main/java/pl/edu/zut/mad/schedule/module/UtilsModule.kt
@@ -4,7 +4,6 @@ import android.app.Application
 import dagger.Module
 import dagger.Provides
 import pl.edu.zut.mad.schedule.util.DatesProvider
-import pl.edu.zut.mad.schedule.util.MessageProviderLogin
 import pl.edu.zut.mad.schedule.util.ModelMapper
 import pl.edu.zut.mad.schedule.util.NetworkConnection
 import javax.inject.Singleton
@@ -19,10 +18,6 @@ internal class UtilsModule {
     @Provides
     @Singleton
     fun provideDatesProvider() = DatesProvider()
-
-    @Provides
-    @Singleton
-    fun provideMessageProviderLogin() = MessageProviderLogin()
 
     @Provides
     @Singleton
