+++ src/main/java/pl/edu/zut/mad/schedule/search/MessageProviderSearch.kt
@@ -1,0 +1,9 @@
+package pl.edu.zut.mad.schedule.search
+
+import pl.edu.zut.mad.schedule.R
+import pl.edu.zut.mad.schedule.util.MessageProvider
+
+class MessageProviderSearch : MessageProvider() {
+
+    override fun getNotFoundMessageId() = R.string.error_album_number_not_found
+}
