+++ src/main/java/pl/edu/zut/mad/schedule/search/SearchModule.kt
@@ -3,7 +3,6 @@ package pl.edu.zut.mad.schedule.search
 import dagger.Module
 import dagger.Provides
 import pl.edu.zut.mad.schedule.data.ScheduleService
-import pl.edu.zut.mad.schedule.util.MessageProviderLogin
 import pl.edu.zut.mad.schedule.util.ModelMapper
 import pl.edu.zut.mad.schedule.util.NetworkConnection
 
@@ -13,6 +12,10 @@ internal class SearchModule(private val view: SearchMvp.View) {
     @Search
     @Provides
     fun providePresenter(service: ScheduleService, modelMapper: ModelMapper,
-        connection: NetworkConnection, messageProvider: MessageProviderLogin): SearchMvp.Presenter =
+        connection: NetworkConnection, messageProvider: MessageProviderSearch): SearchMvp.Presenter =
         SearchPresenter(view, service, modelMapper, connection, messageProvider)
+
+    @Search
+    @Provides
+    fun provideMessageProviderSearch() = MessageProviderSearch()
 }
