+++ src/main/java/pl/edu/zut/mad/schedule/util/MessageProviderLogin.kt
@@ -1,19 +1,8 @@
 package pl.edu.zut.mad.schedule.util
 
 import pl.edu.zut.mad.schedule.R
-import retrofit2.HttpException
-import java.net.HttpURLConnection
 
-internal class MessageProviderLogin {
+internal class MessageProviderLogin : MessageProvider() {
 
-    fun getResIdByError(error: Throwable): Int {
-        if (error is HttpException) {
-            when (error.code()) {
-                HttpURLConnection.HTTP_NOT_FOUND -> return R.string.error_album_number_not_found
-                HttpURLConnection.HTTP_UNAVAILABLE -> return R.string.error_service_database_update
-                HttpURLConnection.HTTP_INTERNAL_ERROR -> return R.string.error_service_internal
-            }
-        }
-        return R.string.error_service_unrecognized
-    }
+    override fun getNotFoundMessageId() = R.string.error_album_number_not_found
 }
