+++ src/main/res/values/strings.xml
@@ -9,6 +9,7 @@
     <string name="error_no_internet">No Internet connection</string>
     <string name="error_field_cannot_be_empty">Field cannot be empty</string>
     <string name="error_album_number_not_found">Given album number not exists</string>
+    <string name="error_schedule_is_empty">Schedule for given queries is empty</string>
     <string name="error_service_database_update">Server database is updating, try again within few minutes</string>
     <string name="error_service_internal">Internal server error, try again within few minutes</string>
     <string name="error_service_unrecognized">Unrecognized error, try again within few minutes</string>
